refactor: Split dependency injection
