+++ src/main.rs
@@ -1,3 +1,4 @@
+mod dependency_injection;
 mod domain;
 mod extractors;
 mod infrastructure;
@@ -6,13 +7,8 @@ mod types;
 mod use_case;
 
 use actix_web::{get, middleware::Logger, web, App, HttpResponse, HttpServer, Responder};
-use infrastructure::author_repository::PgAuthorRepository;
-use presentational::{
-    controller::graphql_controller::graphql,
-    graphql::{query::QueryRoot, query_service::QueryServiceImpl, schema::build_schema},
-};
-use sqlx::postgres::PgPoolOptions;
-use use_case::interactor::author::ShowAuthorInteractor;
+use dependency_injection::{dependency_injection, QSI};
+use presentational::controller::graphql_controller::graphql;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
@@ -20,25 +16,9 @@ async fn main() -> std::io::Result<()> {
 
     env_logger::init();
 
-    let db_url = fetch_database_url();
-
-    let pool = PgPoolOptions::new()
-        .max_connections(5)
-        .connect(&db_url)
-        .await
-        .unwrap();
-
     let auth0_config = extractors::Auth0Config::default();
 
-    let author_repository = PgAuthorRepository { pool };
-    let show_author_use_case = ShowAuthorInteractor::new(author_repository);
-    let query_service = QueryServiceImpl {
-        show_author_use_case,
-    };
-    let query = QueryRoot::new(query_service);
-    let schema = build_schema(query);
-
-    type QSI = QueryServiceImpl<ShowAuthorInteractor<PgAuthorRepository>>;
+    let schema = dependency_injection().await;
 
     HttpServer::new(move || {
         App::new()
@@ -67,15 +47,5 @@ fn fetch_port() -> u16 {
             .expect("Failed to parse environment variable PORT."),
         Err(VarError::NotPresent) => panic!("Environment variable PORT is required."),
         Err(VarError::NotUnicode(_)) => panic!("Environment variable PORT is not unicode."),
-    }
-}
-
-fn fetch_database_url() -> String {
-    use std::env::VarError;
-
-    match std::env::var("DATABASE_URL") {
-        Ok(s) => s,
-        Err(VarError::NotPresent) => panic!("Environment variable DATABASE_URL is required."),
-        Err(VarError::NotUnicode(_)) => panic!("Environment variable DATABASE_URL is not unicode."),
     }
 }
